+++ CHANGELOG.md
@@ -23,6 +23,10 @@ Types of changes:
 
 * cTAKES XMI output produces duplicate CUIs due to different codes; by default, only take the first now to more accurately represent CUI count
 
+### Changed
+
+* For extracting text to files, reduced memory footprint by only retaining most recent 10 records.
+
 ### Added
 
 * Function/script to generate comparisons between different methods/approaches using Jaccard similarity